Remove old apply function
