--- src/olms.js
@@ -1,23 +1,0 @@
-import olms, {
-  apply,
-  applyBackground,
-  applyStyle,
-  getFeatureState,
-  setFeatureState,
-} from './index.js';
-import stylefunction, {
-  recordStyleLayer,
-  renderTransparent,
-} from './stylefunction.js';
-
-export {
-  olms as default,
-  apply,
-  applyBackground,
-  applyStyle,
-  setFeatureState,
-  getFeatureState,
-  stylefunction,
-  recordStyleLayer,
-  renderTransparent,
-};
